WTE-126 Refactor: Create custom hooks for meal and cusine feeds. Refactor Details page.
