+++ src/functions/Feed/index.js
@@ -1,6 +1,9 @@
 import useFetchPopularFeed from './popularFeed/index';
-import useFetchMealFeed from './mealFeed/index';
-import useFetchCuisineFeed from './cuisineFeed/index';
+import { useFetchMealFeed, useFetchMealFeedByTitle } from './mealFeed/index';
+import {
+  useFetchCuisineFeed,
+  useFetchCuisineFeedByTitle,
+} from './cuisineFeed/index';
 import useFetchHealthyFeed from './healthyFeed/index';
 import useFetchRandomFeed from './randomFeed/index';
 import useFetchFeaturedFeed from './featuredFeed/index';
@@ -9,7 +12,9 @@ import useRefreshControl from './refreshControl/index';
 export {
   useFetchPopularFeed,
   useFetchMealFeed,
+  useFetchMealFeedByTitle,
   useFetchCuisineFeed,
+  useFetchCuisineFeedByTitle,
   useFetchHealthyFeed,
   useFetchRandomFeed,
   useFetchFeaturedFeed,
