+++ src/screens/SearchResults.js
@@ -1,10 +1,11 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { StyleSheet, FlatList, SafeAreaView } from 'react-native';
 import { useTheme } from 'react-native-paper';
-import { useSelector, useDispatch } from 'react-redux';
 
-import { handleFetchMealFeedByTitle } from '../features/mealFeed/index';
-import { handleFetchCuisineFeedByTitle } from '../features/cuisineFeed/index';
+import {
+  useFetchMealFeedByTitle,
+  useFetchCuisineFeedByTitle,
+} from '../functions/Feed/index';
 import { SearchResultFeed } from './index';
 import { SearchResultSkeletonCard } from '../common/components/index';
 import { handleSearchResultFeedPlaceholder } from '../utility/index';
@@ -21,24 +22,24 @@ const SearchResults = ({
   randomFeed,
 }) => {
   const { fonts } = useTheme();
-  const dispatch = useDispatch();
-  const { mealFeed, isMealFeedLoaded } = useSelector((state) => state.mealFeed);
-  const { cuisineFeed, isCuisineFeedLoaded } = useSelector(
-    (state) => state.cuisineFeed,
-  );
   const { title, screen } = route ? route.params : {};
+  /**
+   *
+   */
+  const { cuisineFeed, isCuisineFeedLoaded } = useFetchCuisineFeedByTitle(
+    title,
+    screen,
+  );
+  /**
+   *
+   */
+  const { mealFeed, isMealFeedLoaded } = useFetchMealFeedByTitle(title, screen);
+
   /**
    * NOTE: this will need to handle the searched image results
    * of either the searched recipe OR searched meal type / category
    */
-  useEffect(() => {
-    if (title && screen === 'mealFeed') {
-      dispatch(handleFetchMealFeedByTitle(title));
-    }
-    if (title && screen === 'cuisineFeed') {
-      dispatch(handleFetchCuisineFeedByTitle(title));
-    }
-  }, [dispatch, title, screen]);
+
   /**
    *
    * @param {string} feedName
